requirements.txt
aiohappyeyeballs==2.4.6
aiohttp==3.11.12
aiohttp-retry==2.9.1
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.8.0
APScheduler==3.10.4
asgiref==3.5.0
async-timeout==4.0.3
attrs==25.1.0
blinker==1.9.0
certifi==2025.1.31
chardet==4.0.0
charset-normalizer==3.4.1
click==8.1.8
colorhash==2.0.0
configparser==7.1.0
cycler==0.10.0
dataclasses-json==0.6.7
distro==1.9.0
Django==4.0.4
exceptiongroup==1.2.2
Flask==2.0.1
Flask-MonitoringDashboard==3.3.2
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==2.10
importlib_metadata==8.6.1
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
kiwisolver==1.3.1
langchain==0.3.18
langchain-community==0.3.17
langchain-core==0.3.34
langchain-openai==0.3.4
langchain-text-splitters==0.3.6
langsmith==0.3.6
MarkupSafe==3.0.2
marshmallow==3.26.1
matplotlib==3.4.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.63.0
orjson==3.10.15
packaging==24.2
Pillow==8.2.0
propcache==0.2.1
psutil==6.1.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
pyparsing==2.4.7
PyPDF2==3.0.1
python-dateutil==2.8.1
python-dotenv==1.0.1
python-telegram-bot==21.10
pytz==2025.1
PyYAML==6.0.2
regex==2024.11.6
requests==2.25.1
requests-toolbelt==1.0.0
schedule==1.2.2
scipy==1.13.1
six==1.15.0
sniffio==1.3.1
SQLAlchemy==2.0.38
sqlparse==0.4.2
tenacity==9.0.0
tiktoken==0.8.0
tqdm==4.67.1
twilio==9.4.4
typing-inspect==0.9.0
typing_extensions==4.12.2
tzlocal==2.0.0
urllib3==1.26.20
Werkzeug==2.2.2
yarl==1.18.3
zipp==3.21.0
zstandard==0.23.0
